line notes: update mock location
