+++ test/test_helper.exs
@@ -1,1 +1,3 @@
+Code.load_file("test/graph_mock.ex")
+
 ExUnit.start()
